split up js, add setup and build scripts
