+++ js/entry.js
@@ -1,0 +1,491 @@
+
+/**
+ * BEGIN ARGS
+ */
+
+var default_args = {
+    'tuning': ['E', 'A', 'D', 'G', 'B', 'E'],
+    'frets': 15,
+    'start': 0,
+    'end': null,
+    'print_notes': true,
+    'print_numbers': true,
+    
+    'colors': {
+        0: [default_note_fg, default_note_bg],
+        1: [default_note_fg, default_note_bg],
+        2: [default_note_fg, default_note_bg],
+        3: [default_note_fg, default_note_bg],
+        4: [default_note_fg, default_note_bg],
+        5: [default_note_fg, default_note_bg],
+        6: [default_note_fg, default_note_bg],
+        7: [default_note_fg, default_note_bg],
+        8: [default_note_fg, default_note_bg],
+        9: [default_note_fg, default_note_bg],
+       10: [default_note_fg, default_note_bg],
+       11: [default_note_fg, default_note_bg],
+       12: [default_note_fg, default_note_bg]
+    },
+    'scale': {
+        'root': null,
+        'major_formula': null,
+        'chromatic_formula': null,
+        'name': null,
+        'subset': null,
+        'intervals': null,
+        'recolor_intervals': false 
+    },
+    'inlay': {
+        'pattern': null,
+        'color': ['', '']
+    }
+}
+
+default_scale = {
+    'root': 'A',
+    'chromatic_formula': [1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1],
+    'major_formula': ['1', '2', '3', '4', '5', '6', '7'],
+    'name': ['major', 'ionian'],
+    'subset': [1, 3, 5],
+    'intervals': [1, 2, 3, 4, 5, 6, 7] 
+}
+
+function test() {
+    var testStr = ''
+    args = JSON.parse(JSON.stringify(default_args))
+    args.scale.root = 'A'
+    args.scale.name = ['major', 'aeolian']
+    args.colors[1] = ['red', 'white']
+    args.colors[3] = [default_fg, 'light-grey']
+    args.colors[5] = [default_fg, 'grey']
+    args.inlay.pattern = '*'
+    args.inlay.color = [default_fg, '']
+    args.scale.subset = [1, 3, 5]
+    args.scale.intervals = [1, 2, 3, 4, 5, 6, 7]
+    end = args.end == null ? args.frets : args.end
+    fretboards = getFretboardsWithName(args)
+    for (i in fretboards) {
+        var name = fretboards[i][0]
+        var fretboard = fretboards[i][1]
+        testStr += reset_code + name +'\n'
+        testStr += fretboard.str(args.start, end, args.print_notes, args.print_numbers) +'\n'
+    }
+    return testStr
+}
+
+//console.log(test())
+
+/**
+ * END ARGS
+ */
+
+/**
+ * BEGIN WEB DISPLAY
+ */
+
+var Convert = require('ansi-to-html');
+var convert = new Convert();
+ 
+function clearSelect(select) {
+    var length = select.options.length;
+    for (i = length-1; i >= 0; i--) {
+      select.options[i] = null;
+    }
+}
+
+function addTuning() {
+    var app = document.getElementById("app");
+    app.appendChild(document.createTextNode("String Tunings: "));
+    // Create an <input> element, set its type and name attributes
+    var input = document.createElement("input");
+    input.type = "text";
+    input.name = "tuning";
+    input.id = "tuning";
+    input.defaultValue = default_args.tuning.join(' ')
+    app.appendChild(input);
+    // Append a line break
+    app.appendChild(document.createElement("br"));
+}
+
+function addFrets() {
+    var app = document.getElementById("app");
+    app.appendChild(document.createTextNode("Frets: "));
+    // Create an <input> element, set its type and name attributes
+    var input = document.createElement("select");
+    input.name = "frets";
+    input.id = "frets";
+    for (var i = 1; i <=25; i++) {
+        var option = document.createElement("option");
+        option.value = i;
+        option.text = i;
+        input.appendChild(option);
+    }
+    input.selectedIndex= default_args.frets-1
+    input.onchange = setStartEndFromFrets
+    app.appendChild(input);
+    // Append a line break
+    app.appendChild(document.createElement("br"));
+}
+
+function setStartFromFrets() {
+    start = document.getElementById('start')
+    oldStart = start.selectedIndex < 0 ? 0 : start.selectedIndex
+    clearSelect(start)
+    var maxStart = document.getElementById('frets').value - 1
+    for (var i = 0; i <= maxStart; i++) {
+        var option = document.createElement("option");
+        option.value = i;
+        option.text = i;
+        start.appendChild(option);
+    }
+    start.selectedIndex = oldStart > maxStart ? maxStart : oldStart
+}
+
+function setEndFromFrets() {
+    end = document.getElementById('end')
+    var maxEnd = document.getElementById('frets').value
+    var minEnd = parseInt(document.getElementById('start').value) + 1
+    clearSelect(end)
+    for (var i = minEnd; i <= maxEnd; i++) {
+        var option = document.createElement("option");
+        option.value = i;
+        option.text = i;
+        end.appendChild(option);
+    }
+    end.selectedIndex = maxEnd-minEnd
+}
+
+function setStartEndFromFrets() {
+    setStartFromFrets()
+    setEndFromFrets()
+}
+
+function addStart() {
+    var app = document.getElementById("app");
+    app.appendChild(document.createTextNode("Start Fret: "));
+    // Create an <input> element, set its type and name attributes
+    var input = document.createElement("select");
+    input.name = "start";
+    input.id = "start";
+    input.onchange = setEndFromFrets
+    app.appendChild(input);
+    setStartFromFrets()
+    // Append a line break
+    app.appendChild(document.createElement("br"));
+}
+
+function addEnd() {
+    var app = document.getElementById("app");
+    app.appendChild(document.createTextNode("End Fret: "));
+    // Create an <input> element, set its type and name attributes
+    var input = document.createElement("select");
+    input.name = "end";
+    input.id = "end";
+    app.appendChild(input);
+    setEndFromFrets()
+    // Append a line break
+    app.appendChild(document.createElement("br"));
+}
+
+function addPrintNumbers() {
+    var app = document.getElementById("app");
+    app.appendChild(document.createTextNode("Print Note Numbers: "));
+    // Create an <input> element, set its type and name attributes
+    var input = document.createElement("input");
+    input.type = "checkbox"
+    input.id = "print_numbers";
+    input.checked = true
+    app.appendChild(input);
+    // Append a line break
+    app.appendChild(document.createElement("br"));
+}
+
+function addPrintNotes() {
+    var app = document.getElementById("app");
+    app.appendChild(document.createTextNode("Print Note Letters: "));
+    // Create an <input> element, set its type and name attributes
+    var input = document.createElement("input");
+    input.type = "checkbox"
+    input.id = "print_notes";
+    input.checked = true
+    app.appendChild(input);
+    // Append a line break
+    app.appendChild(document.createElement("br"));
+}
+
+function addColor(i, fg, bg) {
+    var app = document.getElementById("app");
+    app.appendChild(document.createTextNode(`Note ${i}: `));
+    // Create an <input> element, set its type and name attributes
+    var fgin = document.createElement("select");
+    var bgin = document.createElement("select");
+    fgin.name = `color_fg_${i}`;
+    bgin.name = `color_bg_${i}`;
+    fgin.id = `color_fg_${i}`;
+    bgin.id = `color_bg_${i}`;
+    for (var i in colors) {
+        var fg_option = document.createElement("option");
+        var bg_option = document.createElement("option");
+        fg_option.value = i;
+        bg_option.value = i;
+        fg_option.text = i;
+        bg_option.text = i;
+        fgin.appendChild(fg_option);
+        bgin.appendChild(bg_option);
+    }
+    fgin.value = fg
+    bgin.value = bg
+    app.appendChild(fgin);
+    app.appendChild(bgin);
+    // Append a line break
+    app.appendChild(document.createElement("br"));
+}
+
+function addColors() {
+    var app = document.getElementById("app");
+    app.appendChild(document.createTextNode(`Set [foreground, background] color for nth note in scale/interval: `));
+    app.appendChild(document.createElement("br"));
+    for (var i = 0; i <= 12; i++) {
+      var fg = i == 0 ? default_fg : default_note_fg
+      var bg = i == 0 ? default_bg : default_note_bg
+      addColor(i, fg, bg)
+    }
+}
+
+function addOutputArgs() {
+    addTuning()
+    addFrets()
+    addStart()
+    addEnd()
+    addPrintNumbers()
+    addPrintNotes()
+    addColors()
+}
+
+function generateFretboards() {
+    var data = ''
+    args = JSON.parse(JSON.stringify(default_args))
+    args.tuning = removeItemAll(document.getElementById('tuning').value.split(' '), '')
+    args.frets = parseInt(document.getElementById('frets').value)
+    args.start = parseInt(document.getElementById('start').value)
+    args.end = parseInt(document.getElementById('end').value)
+    args.print_numbers = document.getElementById('print_numbers').checked
+    args.print_notes = document.getElementById('print_notes').checked
+
+    for (i = 0; i <= 12; i++) {
+        var fg_color = document.getElementById(`color_fg_${i}`).value
+        var bg_color = document.getElementById(`color_bg_${i}`).value
+        args.colors[i] = [fg_color, bg_color]
+    }
+
+    args.scale.root = document.getElementById('root').value
+    if (document.getElementById('chromatic_selector').checked) {
+        chromatic_binary = []
+        removeItemAll(document.getElementById('chromatic_formula').value.split(' '), '').forEach(function(i) {
+            chromatic_binary.push(parseInt(i))
+        })
+        args.scale.chromatic_formula = chromatic_binary
+    } else if (document.getElementById('major_selector').checked) {
+        args.scale.major_formula = removeItemAll(document.getElementById('major_formula').value.split(' '), '')
+    } else if (document.getElementById('scale_name_selector').checked) {
+        args.scale.name = [document.getElementById('scale_name_select').value, document.getElementById('mode_name_select').value]
+    }
+
+    args.inlay.pattern = '*'
+    args.inlay.color = [default_fg, '']
+    subset = removeItemAll(document.getElementById('subset').value.split(' '), '')
+    intervals = removeItemAll(document.getElementById('intervals').value.split(' '), '')
+    recolor_intervals = document.getElementById('recolor_intervals').checked
+    args.scale.subset = subset.map(numStr => parseInt(numStr))
+    args.scale.intervals = intervals.map(numStr => parseInt(numStr))
+    args.recolor_intervals = recolor_intervals
+    console.log(args)
+    fretboards = getFretboardsWithName(args)
+    for (i in fretboards) {
+        var name = fretboards[i][0]
+        var fretboard = fretboards[i][1]
+        data += reset_code + name +'\n'
+        data += fretboard.str(args.start, args.end, args.print_notes, args.print_numbers) +'\n'
+    }
+    document.getElementById('fretboard').innerHTML = convert.toHtml(data)
+}
+
+function addButtons(){
+    var app = document.getElementById("app");
+    var generate = document.createElement("button");
+    generate.id = 'generate';
+    generate.textContent = 'Generate';
+    generate.onclick = generateFretboards;
+    app.appendChild(generate);
+    // Append a line break
+    app.appendChild(document.createElement("br"));
+}
+
+function addFretboardOutput(){
+    // Container <div> where dynamic content will be placed
+    var app = document.getElementById("app");
+    app.appendChild(document.createTextNode("Fretboard output: "));
+    app.appendChild(document.createElement("br"));
+    var fretboardOutput = document.createElement("pre");
+    fretboardOutput.id = "fretboard";
+    app.appendChild(fretboardOutput);
+    // Append a line break
+    app.appendChild(document.createElement("br"));
+}
+
+function addRoot() {
+    var app = document.getElementById("app");
+    app.appendChild(document.createTextNode("Scale Root Note: "));
+    // Create an <input> element, set its type and name attributes
+    var input = document.createElement("input");
+    input.type = "text";
+    input.name = "root";
+    input.id = "root";
+    input.defaultValue = default_scale.root
+    app.appendChild(input);
+    // Append a line break
+    app.appendChild(document.createElement("br"));
+}
+
+function addChromaticFormula() {
+    var app = document.getElementById("app");
+    app.appendChild(document.createTextNode("Chromatic Scale Formula: "));
+    // Create an <input> element, set its type and name attributes
+    var check = document.createElement("input");
+    check.type = "radio"
+    check.name = "scale_selector";
+    check.id = "chromatic_selector";
+    check.checked = true
+
+    var input = document.createElement("input");
+    input.type = "text"
+    input.name = "chromatic_formula";
+    input.id = "chromatic_formula";
+    input.value = default_scale.chromatic_formula.join(' ')
+
+    app.appendChild(check);
+    app.appendChild(input);
+    // Append a line break
+    app.appendChild(document.createElement("br"));
+}
+
+function addMajorFormula() {
+    var app = document.getElementById("app");
+    app.appendChild(document.createTextNode("Major Scale Formula: "));
+    // Create an <input> element, set its type and name attributes
+    var check = document.createElement("input");
+    check.type = "radio"
+    check.name = "scale_selector";
+    check.id = "major_selector";
+    check.checked = false 
+
+    var input = document.createElement("input");
+    input.type = "text"
+    input.name = "major_formula";
+    input.id = "major_formula";
+    input.value = default_scale.major_formula.join(' ')
+
+    app.appendChild(check);
+    app.appendChild(input);
+    // Append a line break
+    app.appendChild(document.createElement("br"));
+}
+
+function setScaleBox() {
+    var scaleBox = document.getElementById(`scale_name_select`)
+    clearSelect(scaleBox)
+    for (var scale in scales) {
+        var option = document.createElement("option");
+        option.value = scale;
+        option.text = scale;
+        scaleBox.appendChild(option)
+    }
+}
+
+function setModeBox() {
+    var modeBox = document.getElementById(`mode_name_select`)
+    var scale = document.getElementById(`scale_name_select`).value
+    clearSelect(modeBox)
+    for (var mode in scales[scale]) {
+        var option = document.createElement("option");
+        option.value = mode;
+        option.text = mode;
+        modeBox.appendChild(option)
+    }
+}
+
+function addScaleName() {
+    var app = document.getElementById("app");
+    app.appendChild(document.createTextNode("Scale/Mode Name: "));
+    // Create an <input> element, set its type and name attributes
+    var check = document.createElement("input");
+    check.type = "radio"
+    check.name = "scale_selector";
+    check.id = "scale_name_selector";
+    check.checked = false
+
+    var scaleBox = document.createElement("select");
+    var modeBox = document.createElement("select");
+    scaleBox.name = `scale_name_select`;
+    modeBox.name = `mode_name_select`;
+    scaleBox.id = `scale_name_select`;
+    modeBox.id = `mode_name_select`;
+    scaleBox.onchange = setModeBox
+
+    app.appendChild(check);
+    app.appendChild(scaleBox);
+    app.appendChild(modeBox);
+    // Append a line break
+    app.appendChild(document.createElement("br"));
+    setScaleBox()
+    setModeBox()
+}
+
+function addSubsets() {
+    var app = document.getElementById("app");
+    // Create an <input> element, set its type and name attributes
+    var subset = document.createElement("input");
+    var intervals = document.createElement("input");
+    subset.type = "text";
+    subset.name = "subset";
+    subset.id = "subset";
+    subset.defaultValue = default_scale.subset.join(' ')
+    intervals.type = "text";
+    intervals.name = "intervals";
+    intervals.id = "intervals";
+    intervals.defaultValue = default_scale.intervals.join(' ')
+    app.appendChild(document.createTextNode("Subset: "));
+    app.appendChild(subset);
+    app.appendChild(document.createTextNode("Intervals: "));
+    app.appendChild(intervals);
+    // Append a line break
+    app.appendChild(document.createElement("br"));
+
+}
+
+function addRecolorSubsets() {
+    var app = document.getElementById("app");
+    app.appendChild(document.createTextNode("Recolor Intervals: "));
+    // Create an <input> element, set its type and name attributes
+    var input = document.createElement("input");
+    input.type = "checkbox"
+    input.id = "recolor_intervals";
+    input.checked = true
+    app.appendChild(input);
+    // Append a line break
+    app.appendChild(document.createElement("br"));
+}
+
+function addScaleSelection() {
+    addRoot()
+    addChromaticFormula()
+    addMajorFormula()
+    addScaleName()
+    addSubsets()
+    addRecolorSubsets()
+}
+
+addOutputArgs()
+addScaleSelection()
+addButtons()
+addFretboardOutput()
