+++ package.json
@@ -18,5 +18,6 @@
   "homepage": "https://github.com/necrosato/fretboard#readme",
   "dependencies": {
     "ansi-to-html": "^0.6.14"
-  }
+  },
+  "devDependencies": {}
 }
