+++ setup.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+npm i -g browserify
+npm i -g uglify-js
+
